+++ package.json
@@ -23,7 +23,8 @@
         "react-dom": "^18.2.0",
         "react-hook-form": "^7.49.3",
         "react-query": "^3.39.3",
-        "react-router-dom": "^6.21.3"
+        "react-router-dom": "^6.21.3",
+        "recoil": "^0.7.7"
     },
     "devDependencies": {
         "@types/react": "^18.2.43",
